Moving Exception classes to their own package
