+++ msal/src/androidTest/java/com/microsoft/identity/client/ClientInfoTest.java
@@ -25,6 +25,9 @@ package com.microsoft.identity.client;
 
 import android.support.test.runner.AndroidJUnit4;
 
+import com.microsoft.identity.client.exception.MsalClientException;
+import com.microsoft.identity.client.exception.MsalException;
+
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
