+++ msal/src/main/java/com/microsoft/identity/client/AuthenticationCallback.java
@@ -23,6 +23,11 @@
 
 package com.microsoft.identity.client;
 
+import com.microsoft.identity.client.exception.MsalClientException;
+import com.microsoft.identity.client.exception.MsalException;
+import com.microsoft.identity.client.exception.MsalServiceException;
+import com.microsoft.identity.client.exception.MsalUiRequiredException;
+
 /**
  * Callback passed with token acquisition. {@link AuthenticationResult} or {@link MsalException} will be returned back via callback.
  */
