+++ msal/src/main/java/com/microsoft/identity/client/exception/MsalArgumentException.java
@@ -1,4 +1,4 @@
-package com.microsoft.identity.client;
+package com.microsoft.identity.client.exception;
 
 public class MsalArgumentException extends MsalException {
