+++ msal/src/main/java/com/microsoft/identity/client/exception/MsalUserCancelException.java
@@ -21,7 +21,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
 
-package com.microsoft.identity.client;
+package com.microsoft.identity.client.exception;
 
 /**
  * MSAL internal exception for user cancelling the flow.
