+++ msal/src/main/java/com/microsoft/identity/client/internal/controllers/MSALTokenCommand.java
@@ -27,8 +27,8 @@ import android.content.Intent;
 import android.support.annotation.NonNull;
 
 import com.microsoft.identity.client.AuthenticationCallback;
-import com.microsoft.identity.client.MsalArgumentException;
-import com.microsoft.identity.client.MsalClientException;
+import com.microsoft.identity.client.exception.MsalArgumentException;
+import com.microsoft.identity.client.exception.MsalClientException;
 import com.microsoft.identity.common.exception.ClientException;
 
 import java.io.IOException;
