+++ msal/src/main/java/com/microsoft/identity/client/internal/controllers/MSALTokenOperation.java
@@ -24,8 +24,8 @@ package com.microsoft.identity.client.internal.controllers;
 
 import android.content.Intent;
 
-import com.microsoft.identity.client.MsalArgumentException;
-import com.microsoft.identity.client.MsalClientException;
+import com.microsoft.identity.client.exception.MsalArgumentException;
+import com.microsoft.identity.client.exception.MsalClientException;
 import com.microsoft.identity.common.exception.ClientException;
 
 import java.io.IOException;
