+++ testapps/testapp/src/main/java/com/microsoft/identity/client/testapp/MainActivity.java
@@ -47,11 +47,11 @@ import com.microsoft.identity.client.IAccount;
 import com.microsoft.identity.client.ILoggerCallback;
 import com.microsoft.identity.client.IMsalEventReceiver;
 import com.microsoft.identity.client.Logger;
-import com.microsoft.identity.client.MsalArgumentException;
-import com.microsoft.identity.client.MsalClientException;
-import com.microsoft.identity.client.MsalException;
-import com.microsoft.identity.client.MsalServiceException;
-import com.microsoft.identity.client.MsalUiRequiredException;
+import com.microsoft.identity.client.exception.MsalArgumentException;
+import com.microsoft.identity.client.exception.MsalClientException;
+import com.microsoft.identity.client.exception.MsalException;
+import com.microsoft.identity.client.exception.MsalServiceException;
+import com.microsoft.identity.client.exception.MsalUiRequiredException;
 import com.microsoft.identity.client.PublicClientApplication;
 import com.microsoft.identity.client.Telemetry;
 import com.microsoft.identity.client.UiBehavior;
